+++ test/unit/search/search_test.cpp
@@ -111,6 +111,13 @@ TYPED_TEST(search_test, invalid_error_configuration)
     EXPECT_THROW(search("A"_dna4, this->index, cfg), std::invalid_argument);
 }
 
+TYPED_TEST(search_test, invalid_dynamic_hit_configuration)
+{
+    seqan3::configuration const cfg = seqan3::search_cfg::max_error{seqan3::search_cfg::total{1}} |
+                                      seqan3::search_cfg::hit{};
+    EXPECT_THROW(search("A"_dna4, this->index, cfg), std::invalid_argument);
+}
+
 TYPED_TEST(search_test, error_substitution)
 {
     {
@@ -211,7 +218,8 @@ TYPED_TEST(search_test, error_levenshtein)
 
     {
         seqan3::configuration const cfg = seqan3::search_cfg::max_error{seqan3::search_cfg::total{2}};
-        EXPECT_RANGE_EQ(search("CCGT"_dna4, this->index, cfg) | position, (std::vector{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10}));
+        EXPECT_RANGE_EQ(search("CCGT"_dna4, this->index, cfg) | position,
+                        (std::vector{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10}));
     }
 }
 
@@ -244,6 +252,12 @@ TYPED_TEST(search_test, search_strategy_all)
     {
         seqan3::configuration const cfg = seqan3::search_cfg::max_error{seqan3::search_cfg::total{1}} |
                                           seqan3::search_cfg::hit_all;
+        EXPECT_RANGE_EQ(search("ACGT"_dna4, this->index, cfg) | position, (std::vector{0, 1, 4, 5, 8, 9}));
+    }
+
+    {
+        seqan3::configuration const cfg = seqan3::search_cfg::max_error{seqan3::search_cfg::total{1}} |
+                                          seqan3::search_cfg::hit{seqan3::search_cfg::hit_all};
         EXPECT_RANGE_EQ(search("ACGT"_dna4, this->index, cfg) | position, (std::vector{0, 1, 4, 5, 8, 9}));
     }
 }
@@ -326,6 +340,16 @@ TYPED_TEST(search_test, search_strategy_best)
         ASSERT_EQ(result.size(), 1u);
         EXPECT_TRUE(std::find(possible_hits2d.begin(), possible_hits2d.end(), result[0]) != possible_hits2d.end());
     }
+
+    {  // Find best match with 2 deletions.
+        std::vector possible_hits2d{0, 4}; // any of 0, 4 ... 1, 5 are not best hits
+        seqan3::configuration const cfg = seqan3::search_cfg::max_error{seqan3::search_cfg::deletion{2}} |
+                                          seqan3::search_cfg::hit{seqan3::search_cfg::hit_single_best};
+
+        std::vector result = search("AGTAGT"_dna4, this->index, cfg) | position | seqan3::views::to<std::vector>;
+        ASSERT_EQ(result.size(), 1u);
+        EXPECT_TRUE(std::find(possible_hits2d.begin(), possible_hits2d.end(), result[0]) != possible_hits2d.end());
+    }
 }
 
 TYPED_TEST(search_test, search_strategy_all_best)
@@ -334,7 +358,18 @@ TYPED_TEST(search_test, search_strategy_all_best)
         seqan3::configuration const cfg = seqan3::search_cfg::max_error{seqan3::search_cfg::total{1}} |
                                           seqan3::search_cfg::hit_all_best;
 
-        EXPECT_RANGE_EQ(search("ACGT"_dna4, this->index, cfg) | position, (std::vector{0, 4, 8})); // 1, 5, 9 are not best hits
+        EXPECT_RANGE_EQ(search("ACGT"_dna4, this->index, cfg) | position,
+                        (std::vector{0, 4, 8})); // 1, 5, 9 are not best hits
+
+        EXPECT_RANGE_EQ(search("AAAA"_dna4, this->index, cfg) | position, (std::vector<int>{})); // no hit
+    }
+
+    {
+        seqan3::configuration const cfg = seqan3::search_cfg::max_error{seqan3::search_cfg::total{1}} |
+                                          seqan3::search_cfg::hit{seqan3::search_cfg::hit_all_best};
+
+        EXPECT_RANGE_EQ(search("ACGT"_dna4, this->index, cfg) | position,
+                        (std::vector{0, 4, 8})); // 1, 5, 9 are not best hits
 
         EXPECT_RANGE_EQ(search("AAAA"_dna4, this->index, cfg) | position, (std::vector<int>{})); // no hit
     }
@@ -357,6 +392,12 @@ TYPED_TEST(search_test, search_strategy_strata)
     {
         seqan3::configuration const cfg = seqan3::search_cfg::max_error{seqan3::search_cfg::total{1}} |
                                           seqan3::search_cfg::hit_strata{1};
+        EXPECT_RANGE_EQ(search("AAAA"_dna4, this->index, cfg) | position, (std::vector<int>{})); // no hit
+    }
+
+    {
+        seqan3::configuration const cfg = seqan3::search_cfg::max_error{seqan3::search_cfg::total{1}} |
+                                          seqan3::search_cfg::hit{seqan3::search_cfg::hit_strata{1}};
         EXPECT_RANGE_EQ(search("AAAA"_dna4, this->index, cfg) | position, (std::vector<int>{})); // no hit
     }